FunController refactoring
